asV0 in-place, rework class splits
