+++ packages/types/src/Metadata/Metadata.spec.ts
@@ -14,41 +14,40 @@ import rpcData from './static';
 describe('Metadata', () => {
   const metadata = new Metadata(rpcData);
 
-  it('decodes latest properly', () => {
-    const str = JSON.stringify(metadata.toJSON());
+  describe('encoding and dceoding', () => {
+    it('decodes latest properly', () => {
+      const str = JSON.stringify(metadata.toJSON());
 
-    console.error(str);
-    console.error(metadata.getUniqTypes());
+      console.error(str);
+      console.error(metadata.getUniqTypes());
 
-    expect(metadata.version).toBe(1);
-    expect(metadata.asV1.modules.length).not.toBe(0);
-    expect(str).toEqual(JSON.stringify(latestParsed));
+      expect(metadata.version).toBe(1);
+      expect(metadata.asV1.modules.length).not.toBe(0);
+      expect(str).toEqual(JSON.stringify(latestParsed));
+    });
+
+    it('converts v1 to V0', () => {
+      const v0 = metadata.asV0;
+
+      // console.error(JSON.stringify(v0.toJSON()));
+
+      expect(metadata.getUniqTypes()).toEqual(v0.getUniqTypes());
+    });
   });
 
   describe('storage with default values', () => {
     Method.injectMethods(extrinsicsFromMeta(metadata.asV0));
 
-    metadata.asV1.modules.forEach((mod) => {
-      if (mod.storage.isNone) {
-        return;
-      }
-
-      mod.storage.unwrap().forEach((fn) => {
-        it(`creates default types for ${mod.prefix}.${fn.name}, type ${fn.type}`, () => {
-          expect(
-            () => createType(fn.type.toString(), fn.default)
-          ).not.toThrow();
-        });
-      });
-    });
-  });
-
-  it.skip('converts v1 to V0', () => {
-    const v0 = metadata.asV0;
-    const str = JSON.stringify(metadata.toJSON());
-
-    console.error(str);
-
-    expect(metadata.getUniqTypes()).toEqual(v0.getUniqTypes());
+    metadata.asV1.modules
+      .filter(({ storage }) => storage.isSome)
+      .map((mod) =>
+        mod.storage.unwrap().forEach((fn) => {
+          it(`creates default types for ${mod.prefix}.${fn.name}, type ${fn.type}`, () => {
+            expect(
+              () => createType(fn.type.toString(), fn.default)
+            ).not.toThrow();
+          });
+        })
+      );
   });
 });
