+++ packages/types/src/Metadata/v1/Events.ts
@@ -1,0 +1,44 @@
+// Copyright 2017-2019 @polkadot/types authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import Struct from '../../codec/Struct';
+import Vector from '../../codec/Vector';
+import Text from '../../Text';
+import Type from '../../Type';
+
+/**
+ * @name MetadataEvent
+ * @description
+ * The definition of an event
+ */
+export class MetadataEvent extends Struct {
+  constructor (value?: any) {
+    super({
+      name: Text,
+      arguments: Vector.with(Type),
+      documentation: Vector.with(Text)
+    }, value);
+  }
+
+  /**
+   * @description The [[Type]] for arguments
+   */
+  get arguments (): Vector<Type> {
+    return this.get('arguments') as Vector<Type>;
+  }
+
+  /**
+   * @description The [[Text]] documentation
+   */
+  get documentation (): Vector<Text> {
+    return this.get('documentation') as Vector<Text>;
+  }
+
+  /**
+   * @description The call name
+   */
+  get name (): Text {
+    return this.get('name') as Text;
+  }
+}
