+++ packages/types/src/Metadata/v1/index.ts
@@ -4,260 +4,15 @@
 
 import { MetadataInterface } from '../types';
 
-import EnumType from '../../codec/EnumType';
 import Option from '../../codec/Option';
 import Struct from '../../codec/Struct';
 import Vector from '../../codec/Vector';
-import Bytes from '../../Bytes';
-import Null from '../../Null';
 import Text from '../../Text';
-import Type from '../../Type';
-import u16 from '../../u16';
 import { flattenUniq, validateTypes } from '../util';
 
-class MetadataCallArg extends Struct {
-  constructor (value?: any) {
-    super({
-      name: Text,
-      type: Type
-    }, value);
-  }
-
-  /**
-   * @description The argument name
-   */
-  get name (): Text {
-    return this.get('name') as Text;
-  }
-
-  /**
-   * @description The [[Type]]
-   */
-  get type (): Type {
-    return this.get('type') as Type;
-  }
-}
-
-/**
- * @name MetadataCall
- * @description
- * The definition of a call
- */
-export class MetadataCall extends Struct {
-  constructor (value?: any) {
-    super({
-      id: u16,
-      name: Text,
-      arguments: Vector.with(MetadataCallArg),
-      documentation: Vector.with(Text)
-    }, value);
-  }
-
-  /**
-   * @description The [[MetadataCallArg]] for arguments
-   */
-  get arguments (): Vector<MetadataCallArg> {
-    return this.get('arguments') as Vector<MetadataCallArg>;
-  }
-
-  /**
-   * @description The [[Text]] documentation
-   */
-  get documentation (): Vector<Text> {
-    return this.get('documentation') as Vector<Text>;
-  }
-
-  /**
-   * @description The call function id
-   */
-  get id (): u16 {
-    return this.get('id') as u16;
-  }
-
-  /**
-   * @description The call name
-   */
-  get name (): Text {
-    return this.get('name') as Text;
-  }
-}
-
-/**
- * @name MetadataEvent
- * @description
- * The definition of an event
- */
-export class MetadataEvent extends Struct {
-  constructor (value?: any) {
-    super({
-      name: Text,
-      arguments: Vector.with(Type),
-      documentation: Vector.with(Text)
-    }, value);
-  }
-
-  /**
-   * @description The [[Type]] for arguments
-   */
-  get arguments (): Vector<Type> {
-    return this.get('arguments') as Vector<Type>;
-  }
-
-  /**
-   * @description The [[Text]] documentation
-   */
-  get documentation (): Vector<Text> {
-    return this.get('documentation') as Vector<Text>;
-  }
-
-  /**
-   * @description The call name
-   */
-  get name (): Text {
-    return this.get('name') as Text;
-  }
-}
-
-class Default extends Null {}
-
-class Optional extends Null {}
-
-class MetadataStorageModifier extends EnumType<Optional | Default> {
-  constructor (value?: any, index?: number) {
-    super({
-      Optional,
-      Default
-    }, value, index);
-  }
-
-  /**
-   * @description `true` if the storage entry is optional
-   */
-  get isOptional (): boolean {
-    return this.toNumber() === 0;
-  }
-
-  toJSON (): any {
-    return this.toString();
-  }
-}
-
-class MapType extends Struct {
-  constructor (value?: any) {
-    super({
-      key: Type,
-      value: Type
-    }, value);
-  }
-
-  /**
-   * @description The mapped key as [[Type]]
-   */
-  get key (): Type {
-    return this.get('key') as Type;
-  }
-
-  /**
-   * @description The mapped value as [[Type]]
-   */
-  get value (): Type {
-    return this.get('value') as Type;
-  }
-}
-
-class PlainType extends Type {
-}
-
-class MetadataStorageType extends EnumType<PlainType | MapType> {
-  constructor (value?: any, index?: number) {
-    super({
-      PlainType,
-      MapType
-    }, value, index);
-  }
-
-  /**
-   * @description `true` if the storage entry is a map
-   */
-  get isMap (): boolean {
-    return this.toNumber() === 1;
-  }
-
-  /**
-   * @description The value as a mapped value
-   */
-  get asMap (): MapType {
-    return this.value as MapType;
-  }
-
-  /**
-   * @description The value as a [[Type]] value
-   */
-  get asType (): PlainType {
-    return this.value as PlainType;
-  }
-
-  /**
-   * @description Returns the string representation of the value
-   */
-  toString (): string {
-    return this.isMap
-      ? this.asMap.value.toString()
-      : this.asType.toString();
-  }
-}
-
-/**
- * @name MetadataModule
- * @description
- * The definition of a storage function
- */
-export class MetadataStorage extends Struct {
-  constructor (value?: any) {
-    super({
-      name: Text,
-      modifier: MetadataStorageModifier,
-      type: MetadataStorageType,
-      default: Bytes,
-      documentation: Vector.with(Text)
-    }, value);
-  }
-
-  /**
-   * @description The [[Bytes]] default
-   */
-  get default (): Bytes {
-    return this.get('default') as Bytes;
-  }
-
-  /**
-   * @description The [[Text]] documentation
-   */
-  get documentation (): Vector<Text> {
-    return this.get('documentation') as Vector<Text>;
-  }
-
-  /**
-   * @description The [[MetadataArgument]] for arguments
-   */
-  get modifier (): MetadataStorageModifier {
-    return this.get('modifier') as MetadataStorageModifier;
-  }
-
-  /**
-   * @description The call name
-   */
-  get name (): Text {
-    return this.get('name') as Text;
-  }
-
-  /**
-   * @description The [[MetadataStorageType]]
-   */
-  get type (): MetadataStorageType {
-    return this.get('type') as MetadataStorageType;
-  }
-}
+import { MetadataCall } from './Calls';
+import { MetadataEvent } from './Events';
+import { MetadataStorage } from './Storage';
 
 /**
  * @name MetadataModule
